display item when list is selected

+++ LAB19-CreateAnAPI/wwwroot/todo.js
@@ -26,12 +26,14 @@ function getData() {
             getCount(data.length);
             $.each(data, function (key, item) {
                 const checked = item.isComplete ? 'checked' : '';
-
-                $('<tr><td><input disabled="true" type="checkbox" ' + checked + '></td>' +
-                    '<td>' + item.name + '</td>' +
-                    '<td><button onclick="editItem(' + item.id + ')">Edit</button></td>' +
-                    '<td><button onclick="deleteItem(' + item.id + ')">Delete</button></td>' +
-                    '</tr>').appendTo($('#todos'));
+                if (item.listId === selectList) {
+                    $('<tr><td><input disabled="true" type="checkbox" ' + checked + '></td>' +
+                        '<td>' + item.name + '</td>' +
+                        '<td><button onclick="editItem(' + item.id + ')">Edit</button></td>' +
+                        '<td><button onclick="deleteItem(' + item.id + ')">Delete</button></td>' +
+                        '</tr>').appendTo($('#todos'));
+                }
+                
             });
 
             todos = data;
@@ -42,6 +44,7 @@ function getData() {
 function addItem() {
     const item = {
         'name': $('#add-name').val(),
+        'ListId': selectList,
         'isComplete': false
     };
